+++ src/Graph/Nodes/Cluster.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\Graph\Nodes;
 
 use Rubix\Tensor\Matrix;
+use Rubix\ML\Datasets\Dataset;
 use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Kernels\Distance\Distance;
 use InvalidArgumentException;
@@ -22,7 +23,7 @@ class Cluster extends BinaryNode implements Ball, Leaf
     /**
      * The samples that make up the cluster.
      *
-     * @var array
+     * @var array[]
      */
     protected $samples;
 
@@ -34,14 +35,14 @@ class Cluster extends BinaryNode implements Ball, Leaf
     protected $labels;
 
     /**
-     * The center or multivariate mean of the cluster.
+     * The centroid or multivariate mean of the cluster.
      *
-     * @var array
+     * @var (int|float)[]
      */
     protected $center;
 
     /**
-     * The radius of the cluster.
+     * The radius of the cluster centroid.
      *
      * @var float
      */
@@ -50,14 +51,14 @@ class Cluster extends BinaryNode implements Ball, Leaf
     /**
      * Terminate a branch with a dataset.
      *
-     * @param \Rubix\ML\Datasets\Labeled $dataset
+     * @param \Rubix\ML\Datasets\Dataset $dataset
      * @param \Rubix\ML\Kernels\Distance\Distance $kernel
      * @return self
      */
-    public static function terminate(Labeled $dataset, Distance $kernel) : self
+    public static function terminate(Dataset $dataset, Distance $kernel) : self
     {
         $samples = $dataset->samples();
-        $labels = $dataset->labels();
+        $labels = $dataset instanceof Labeled ? $dataset->labels() : null;
 
         $center = Matrix::quick($samples)
             ->transpose()
@@ -77,18 +78,18 @@ class Cluster extends BinaryNode implements Ball, Leaf
 
     /**
      * @param array $samples
-     * @param array $labels
+     * @param array|null $labels
      * @param array $center
      * @param float $radius
      * @throws \InvalidArgumentException
      */
-    public function __construct(array $samples, array $labels, array $center, float $radius)
+    public function __construct(array $samples, ?array $labels, array $center, float $radius)
     {
         if (empty($samples)) {
             throw new InvalidArgumentException('Cluster cannot be empty');
         }
 
-        if (count($samples) !== count($labels)) {
+        if (isset($labels) and count($samples) !== count($labels)) {
             throw new InvalidArgumentException('The number of samples'
                 . ' must be equal to the number of labels.');
         }
@@ -104,7 +105,7 @@ class Cluster extends BinaryNode implements Ball, Leaf
         }
 
         $this->samples = $samples;
-        $this->labels = $labels;
+        $this->labels = $labels ?? [];
         $this->center = $center;
         $this->radius = $radius;
     }
@@ -142,9 +143,9 @@ class Cluster extends BinaryNode implements Ball, Leaf
     /**
      * Return the labels in the cluster.
      *
-     * @return array
+     * @return array|null
      */
-    public function labels() : array
+    public function labels() : ?array
     {
         return $this->labels;
     }
@@ -153,10 +154,10 @@ class Cluster extends BinaryNode implements Ball, Leaf
      * Return the label cooresponding to the ith sample in the cluster.
      *
      * @param int $index
-     * @return mixed
+     * @return mixed|null
      */
     public function label(int $index)
     {
-        return $this->labels[$index];
+        return $this->labels[$index] ?? null;
     }
 }