+++ tests/Graph/Nodes/HypersphereTest.php
@@ -4,12 +4,12 @@ namespace Rubix\ML\Tests\Graph\Nodes;
 
 use Rubix\ML\Datasets\Labeled;
 use Rubix\ML\Graph\Nodes\Node;
-use Rubix\ML\Graph\Nodes\Centroid;
+use Rubix\ML\Graph\Nodes\Hypersphere;
 use Rubix\ML\Graph\Nodes\BinaryNode;
 use Rubix\ML\Kernels\Distance\Euclidean;
 use PHPUnit\Framework\TestCase;
 
-class CentroidTest extends TestCase
+class HypersphereTest extends TestCase
 {
     protected const COLUMN = 1;
     protected const VALUE = 3.;
@@ -33,9 +33,9 @@ class CentroidTest extends TestCase
             Labeled::quick([self::SAMPLES[1]], [self::LABELS[1]]),
         ];
 
-        $node = new Centroid(self::CENTER, self::RADIUS, $groups);
+        $node = new Hypersphere(self::CENTER, self::RADIUS, $groups);
 
-        $this->assertInstanceOf(Centroid::class, $node);
+        $this->assertInstanceOf(Hypersphere::class, $node);
         $this->assertInstanceOf(BinaryNode::class, $node);
         $this->assertInstanceOf(Node::class, $node);
 
@@ -48,7 +48,7 @@ class CentroidTest extends TestCase
     {
         $dataset = Labeled::quick(self::SAMPLES, self::LABELS);
 
-        $node = Centroid::split($dataset, new Euclidean());
+        $node = Hypersphere::split($dataset, new Euclidean());
 
         $this->assertEquals(self::CENTER, $node->center());
         $this->assertEquals(self::RADIUS, $node->radius());